# backend/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ResponseMessage": {
				"properties": {
					"message": {
						"type": "string"
					},
					"detalles": {}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ResponseMessageWithData_any_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"detalles": {},
					"data": {}
				},
				"required": [
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"_36_Enums.EstadoProducto": {
				"type": "string",
				"enum": [
					"DISPONIBLE",
					"AGOTADO",
					"RESERVADO",
					"VENCIDO",
					"ELIMINADO"
				]
			},
			"EstadoProducto": {
				"$ref": "#/components/schemas/_36_Enums.EstadoProducto"
			},
			"ProductosDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"codigoBarras": {
						"type": "string",
						"nullable": true
					},
					"codigoQR": {
						"type": "string",
						"nullable": true
					},
					"nombre": {
						"type": "string"
					},
					"descripcion": {
						"type": "string"
					},
					"cantidad": {
						"type": "number",
						"format": "double"
					},
					"precio": {
						"type": "number",
						"format": "double"
					},
					"fechaAdquisicion": {
						"type": "string"
					},
					"fechaVencimiento": {
						"type": "string"
					},
					"usuarioId": {
						"type": "number",
						"format": "double"
					},
					"estado": {
						"$ref": "#/components/schemas/EstadoProducto"
					},
					"imagen": {
						"type": "string"
					},
					"categoria": {
						"type": "string"
					}
				},
				"required": [
					"nombre",
					"descripcion",
					"cantidad",
					"precio",
					"fechaAdquisicion",
					"fechaVencimiento",
					"usuarioId",
					"estado",
					"imagen"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_ProductosDTO_": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"codigoBarras": {
						"type": "string"
					},
					"codigoQR": {
						"type": "string"
					},
					"nombre": {
						"type": "string"
					},
					"descripcion": {
						"type": "string"
					},
					"cantidad": {
						"type": "number",
						"format": "double"
					},
					"precio": {
						"type": "number",
						"format": "double"
					},
					"fechaAdquisicion": {
						"type": "string"
					},
					"fechaVencimiento": {
						"type": "string"
					},
					"usuarioId": {
						"type": "number",
						"format": "double"
					},
					"estado": {
						"$ref": "#/components/schemas/_36_Enums.EstadoProducto"
					},
					"imagen": {
						"type": "string"
					},
					"categoria": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			},
			"legacyJwt": {
				"type": "apiKey",
				"in": "header",
				"name": "Authorization"
			}
		}
	},
	"info": {
		"title": "producttrack",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "@mariaximena0724 y emilia"
		}
	},
	"paths": {
...
		"/Productos": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Productos"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "nombre",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "categoria",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "estado",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "fechaAdquisicionDesde",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "fechaAdquisicionHasta",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "fechaVencimientoDesde",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "fechaVencimientoHasta",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "usuarioId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Producto creado correctamente",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ResponseMessageWithData_any_"
										},
										{
											"$ref": "#/components/schemas/ResponseMessage"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Datos inválidos"
					}
				},
				"tags": [
					"Productos"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProductosDTO"
							}
						}
					}
				}
			}
		},
		"/Productos/categorias": {
			"get": {
				"operationId": "ObtenerCategorias",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"type": "string"
											},
											"type": "array"
										},
										{
											"$ref": "#/components/schemas/ResponseMessage"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Productos"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "tipoUsuario",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/Productos/por-categoria": {
			"get": {
				"operationId": "GetByCategoria",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Productos"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "categoria",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/Productos/nombres/{idUsuario}": {
			"get": {
				"operationId": "GetNombresProductosDelUsuario",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"type": "string"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Productos"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "idUsuario",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/Productos/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Productos"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateProducto",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseMessage"
								}
							}
						}
					}
				},
				"tags": [
					"Productos"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_ProductosDTO_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteProducto",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseMessage"
								}
							}
						}
					}
				},
				"tags": [
					"Productos"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}